+++ ntex/src/http/client/connect.rs
@@ -18,14 +18,14 @@ pub(crate) struct ConnectorWrapper<T>(pub(crate) T);
 
 pub(crate) trait Connect {
     fn send_request(
-        &mut self,
+        &self,
         head: RequestHead,
         body: Body,
         addr: Option<net::SocketAddr>,
     ) -> Pin<Box<dyn Future<Output = Result<ClientResponse, SendRequestError>>>>;
 
     fn send_request_extra(
-        &mut self,
+        &self,
         head: Rc<RequestHead>,
         extra_headers: Option<HeaderMap>,
         body: Body,
@@ -34,7 +34,7 @@ pub(crate) trait Connect {
 
     /// Send request, returns Response and Framed
     fn open_tunnel(
-        &mut self,
+        &self,
         head: RequestHead,
         addr: Option<net::SocketAddr>,
     ) -> Pin<
@@ -50,7 +50,7 @@ pub(crate) trait Connect {
 
     /// Send request and extra headers, returns Response and Framed
     fn open_tunnel_extra(
-        &mut self,
+        &self,
         head: Rc<RequestHead>,
         extra_headers: Option<HeaderMap>,
         addr: Option<net::SocketAddr>,
@@ -76,7 +76,7 @@ where
     T::Future: 'static,
 {
     fn send_request(
-        &mut self,
+        &self,
         head: RequestHead,
         body: Body,
         addr: Option<net::SocketAddr>,
@@ -99,7 +99,7 @@ where
     }
 
     fn send_request_extra(
-        &mut self,
+        &self,
         head: Rc<RequestHead>,
         extra_headers: Option<HeaderMap>,
         body: Body,
@@ -124,7 +124,7 @@ where
     }
 
     fn open_tunnel(
-        &mut self,
+        &self,
         head: RequestHead,
         addr: Option<net::SocketAddr>,
     ) -> Pin<
@@ -156,7 +156,7 @@ where
     }
 
     fn open_tunnel_extra(
-        &mut self,
+        &self,
         head: Rc<RequestHead>,
         extra_headers: Option<HeaderMap>,
         addr: Option<net::SocketAddr>,