+++ ntex/CHANGES.md
@@ -1,5 +1,13 @@
 # Changes
 
+## [0.1.4] - 2020-04-06
+
+* Remove unneeded RefCell from client connector
+
+* Add trace entries for http1 disaptcher
+
+* Properly set timeout for test http client
+
 ## [0.1.3] - 2020-04-06
 
 * Add server ssl handshake timeout